requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
mysql-connector-python>=8.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
Werkzeug>=2.3.0

